Changed name of a class to ShoppingListService and moved it to service directory
