+++ src/main/java/com/infoshareacademy/food_preferences/FoodPreferences.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.service.food_preferences;
+package com.infoshareacademy.food_preferences;
 
 public class FoodPreferences {
     private AllergenName allergenName;
