+++ src/main/java/com/infoshareacademy/food_preferences/Meal.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.service.food_preferences;
+package com.infoshareacademy.food_preferences;
 
 public enum Meal {
     BREAKFAST,
