+++ src/main/java/com/infoshareacademy/recipe/Recipe.java
@@ -1,8 +1,7 @@
-package com.infoshareacademy.service.recipe;
+package com.infoshareacademy.recipe;
 
-import com.infoshareacademy.food_preferences.*;
-import com.infoshareacademy.service.food_preferences.FoodPreferences;
-import com.infoshareacademy.service.product.Product;
+import com.infoshareacademy.product.Product;
+import com.infoshareacademy.food_preferences.FoodPreferences;
 
 import java.util.List;
 import java.util.Map;
